requirements.txt
beautifulsoup4==4.5.1
click==6.6
py-stackexchange==2.2.6
PyGithub==1.29
pymongo==3.3.0
python-dotenv==0.6.0
requests==2.11.1
six==1.10.0
